+++ lib/delayed_job.rb
@@ -1,13 +1,8 @@
+require 'active_support'
 require 'delayed/compatibility'
 require 'delayed/exceptions'
 require 'delayed/message_sending'
 require 'delayed/performable_method'
-
-if defined?(ActionMailer)
-  require 'action_mailer/version'
-  require 'delayed/performable_mailer'
-end
-
 require 'delayed/yaml_ext'
 require 'delayed/lifecycle'
 require 'delayed/plugin'
@@ -18,5 +13,14 @@ require 'delayed/worker'
 require 'delayed/deserialization_error'
 require 'delayed/railtie' if defined?(Rails::Railtie)
 
+ActiveSupport.on_load(:action_mailer) do
+  require 'delayed/performable_mailer'
+  ActionMailer::Base.extend(Delayed::DelayMail)
+end
+
+module Delayed
+  autoload :PerformableMailer, 'delayed/performable_mailer'
+end
+
 Object.send(:include, Delayed::MessageSending)
-Module.send(:include, Delayed::MessageSending::ClassMethods)
+Module.send(:include, Delayed::MessageSendingClassMethods)