+++ CHANGELOG.md
@@ -1,3 +1,43 @@
+4.1.8 - 2019-08-16
+=================
+* Support for Rails 6.0.0
+
+4.1.7 - 2019-06-20
+=================
+* Fix loading Delayed::PerformableMailer when ActionMailer isn't loaded yet
+
+4.1.6 - 2019-06-19
+=================
+* Properly initialize ActionMailer outside railties (#1077)
+* Fix Psych load_tags support (#1093)
+* Replace REMOVED with FAILED in log message (#1048)
+* Misc doc updates (#1052, #1074, #1064, #1063)
+
+4.1.5 - 2018-04-13
+=================
+* Allow Rails 5.2
+
+4.1.4 - 2017-12-29
+=================
+* Use `yaml_tag` instead of deprecated `yaml_as` (#996)
+* Support ruby 2.5.0
+
+4.1.3 - 2017-05-26
+=================
+* Don't mutate the options hash (#877)
+* Log an error message when a deserialization error occurs (#894)
+* Adding the queue name to the log output (#917)
+* Don't include ClassMethods with MessageSending (#924)
+* Fix YAML deserialization error if original object is soft-deleted (#947)
+* Add support for Rails 5.1 (#982)
+
+4.1.2 - 2016-05-16
+==================
+* Added Delayed::Worker.queue_attributes
+* Limit what we require in ActiveSupport
+* Fix pid file creation when there is no tmp directory
+* Rails 5 support
+
 4.1.1 - 2015-09-24
 ==================
 * Fix shared specs for back-ends that reload objects